Add view model to main window, this is final working version
