+++ DocViewer/MainWindow.xaml
@@ -3,12 +3,12 @@
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local1="clr-namespace:DocViewer.ViewModels"
+        xmlns:local="clr-namespace:DocViewer.ViewModels"
         mc:Ignorable="d"
         Title="MainWindow" Height="450" Width="800">
 
     <Window.Resources>
-        <local1:PanelControlPageViewModel x:Key="ABC" />
+        <local:MainWindowViewModel x:Key="ABC" />
     </Window.Resources>
 
     <Grid>
